[#1745] Type definition storage creation is moved out of publish subscribe resources
